pyproject.toml
[tool.poetry]
name = "tms_fin_project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.1.2"
selenium = "^4.3.0"
allure-pytest = "^2.9.45"
fake-useragent = "^0.1.11"
pytest-rerunfailures = "^10.2"
requests = "^2.28.1"
pre-commit="2.16.0"
Faker = "^13.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
